move type definitions to `friendly_traceback.typing`
