+++ friendly_traceback/info_generic.py
@@ -6,9 +6,7 @@ import inspect
 from typing import Any, Callable, Dict, Type
 
 from .ft_gettext import current_lang, no_information
-
-
-GenericExplain = Callable[[], str]
+from .typing import GenericExplain
 
 
 GENERIC: Dict[Type[BaseException], GenericExplain] = {}
